Cleanup: - Moved everything to one package, runfolder for simplicity (there is nothing to be reused in that arteria package) - Necessary cleanup in namespaces etc because of this
