--- dependencies
@@ -1,2 +1,0 @@
-pip install tornado
-pip install jsonpickle
