--- runfolder-ws/arteria
@@ -1,1 +1,0 @@
-../arteria/
